+++ lib/openhab/core/types/type.rb
@@ -3,9 +3,14 @@
 module OpenHAB
   module Core
     module Types
-      java_import org.openhab.core.types.Command,
-                  org.openhab.core.types.Type,
-                  org.openhab.core.types.State
+      # @interface
+      Type = org.openhab.core.types.Type
+
+      # @interface
+      Command = org.openhab.core.types.Command
+
+      # @interface
+      State = org.openhab.core.types.State
 
       # This is a parent interface for all {State}s and {Command}s. It
       # was introduced as many states can be commands at the same time and vice
@@ -92,15 +97,17 @@ module OpenHAB
         end
       end
 
-      # This is a marker interface for all command types.
-      module Command
-        # @!parse include Type
-      end
-
-      # This is a marker interface for all state types.
-      module State
-        # @!parse include Type
-      end
+      #
+      # @!parse
+      #   # This is a marker interface for all command types.
+      #   module Command
+      #     include Type
+      #   end
+      #
+      #   # This is a marker interface for all state types.
+      #   module State
+      #     include Type
+      #   end
     end
   end
 end